+++ src/main/java/searchengine/services/search/SearchService.java
@@ -3,5 +3,5 @@ package searchengine.services.search;
 import searchengine.dto.search.SearchResponse;
 
 public interface SearchService {
-    SearchResponse search(String query, String site, Long offset, Long limit);
+    SearchResponse search(String query, String site, Integer offset, Integer limit);
 }
